+++ test/Resque/Tests/EventTest.php
@@ -1,7 +1,8 @@
 <?php
 use Resque\Resque;
+use Resque\Event;
 /**
- * Resque_Event tests.
+ * Event tests.
  *
  * @package     Resque/Tests
  * @author      Chris Boulton <chris@bigcommerce.com>
@@ -23,7 +24,7 @@ class Resque_Tests_EventTest extends Resque_Tests_TestCase
 
     public function tearDown()
     {
-        Resque_Event::clearListeners();
+        Event::clearListeners();
         $this->callbacksHit = array();
     }
 
@@ -54,7 +55,7 @@ class Resque_Tests_EventTest extends Resque_Tests_TestCase
      */
     public function testEventCallbacksFire($event, $callback)
     {
-        Resque_Event::listen($event, array($this, $callback));
+        Event::listen($event, array($this, $callback));
 
         $job = $this->getEventTestJob();
         $this->worker->perform($job);
@@ -68,7 +69,7 @@ class Resque_Tests_EventTest extends Resque_Tests_TestCase
         $event = 'beforeFork';
         $callback = 'beforeForkEventCallback';
 
-        Resque_Event::listen($event, array($this, $callback));
+        Event::listen($event, array($this, $callback));
         Resque::enqueue('jobs', 'Test_Job', array(
             'somevar'
         ));
@@ -82,7 +83,7 @@ class Resque_Tests_EventTest extends Resque_Tests_TestCase
         $event = 'beforeEnqueue';
         $callback = 'beforeEnqueueEventCallback';
 
-        Resque_Event::listen($event, array($this, $callback));
+        Event::listen($event, array($this, $callback));
         Resque::enqueue('jobs', 'Test_Job', array(
             'somevar'
         ));
@@ -92,7 +93,7 @@ class Resque_Tests_EventTest extends Resque_Tests_TestCase
     public function testBeforePerformEventCanStopWork()
     {
         $callback = 'beforePerformEventDontPerformCallback';
-        Resque_Event::listen('beforePerform', array($this, $callback));
+        Event::listen('beforePerform', array($this, $callback));
 
         $job = $this->getEventTestJob();
 
@@ -104,8 +105,8 @@ class Resque_Tests_EventTest extends Resque_Tests_TestCase
     public function testBeforeEnqueueEventStopsJobCreation()
     {
         $callback = 'beforeEnqueueEventDontCreateCallback';
-        Resque_Event::listen('beforeEnqueue', array($this, $callback));
-        Resque_Event::listen('afterEnqueue', array($this, 'afterEnqueueEventCallback'));
+        Event::listen('beforeEnqueue', array($this, $callback));
+        Event::listen('afterEnqueue', array($this, 'afterEnqueueEventCallback'));
 
         $result = Resque::enqueue('test_job', 'TestClass');
         $this->assertContains($callback, $this->callbacksHit, $callback . ' callback was not called');
@@ -118,7 +119,7 @@ class Resque_Tests_EventTest extends Resque_Tests_TestCase
         $callback = 'afterEnqueueEventCallback';
         $event    = 'afterEnqueue';
 
-        Resque_Event::listen($event, array($this, $callback));
+        Event::listen($event, array($this, $callback));
         Resque::enqueue('jobs', 'Test_Job', array(
             'somevar'
         ));
@@ -130,15 +131,15 @@ class Resque_Tests_EventTest extends Resque_Tests_TestCase
         $callback = 'beforePerformEventCallback';
         $event    = 'beforePerform';
 
-        Resque_Event::listen($event, array($this, $callback));
-        Resque_Event::stopListening($event, array($this, $callback));
+        Event::listen($event, array($this, $callback));
+        Event::stopListening($event, array($this, $callback));
 
         $job = $this->getEventTestJob();
         $this->worker->perform($job);
         $this->worker->work(0);
 
         $this->assertNotContains($callback, $this->callbacksHit,
-            $event . ' callback (' . $callback .') was called though Resque_Event::stopListening was called'
+            $event . ' callback (' . $callback .') was called though Event::stopListening was called'
         );
     }