+++ CHANGELOG.md
@@ -52,7 +52,7 @@ Changes by iskandar introduce improved support for using DSNs to connect to Redi
 * Allow UNIX socket to be passed to Resque when connecting to Redis (pedroarnal)
 * Fix typographical errors in PHP docblocks (chaitanyakuber)
 * Set the queue name on job instances when jobs are executed (chaitanyakuber)
-* Fix and add tests for Resque_Event::stopListening (ebernhardson)
+* Fix and add tests for Event::stopListening (ebernhardson)
 * Documentation cleanup (maetl)
 * Pass queue name to afterEvent callback
 * Only declare RedisException if it doesn't already exist (Matt Heath)